Revert "Added Unit tests for TextDisplay. Altered TextDisplay to allow unit tests to simulate keypress. Fixed Backstory cutscene bug"
This reverts commit de665b0b9d8fe38feaf04b603214abfb739c68b8.
